How to quickly spot relationships in a dataset using Seaborn's .pairplot().
How to split the data into a training and testing dataset to better evaluate a model's performance.

How to run a multivariable regression.

How to evaluate that regression-based on the sign of its coefficients.

How to analyse and look for patterns in a model's residuals.

How to improve a regression model using (a log) data transformation.

How to specify your own values for various features and use your model to make a prediction.